--- workspaces/orchestrator/plugins/scaffolder-backend-module-orchestrator/src/actions/example.ts
@@ -1,60 +1,0 @@
-/*
- * Copyright Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
-
-/**
- * Creates an `acme:example` Scaffolder action.
- *
- * @remarks
- *
- * See {@link https://example.com} for more information.
- *
- * @public
- */
-export function createExampleAction() {
-  // For more information on how to define custom actions, see
-  //   https://backstage.io/docs/features/software-templates/writing-custom-actions
-  return createTemplateAction<{
-    myParameter: string;
-  }>({
-    id: 'acme:example',
-    description: 'Runs an example action',
-    schema: {
-      input: {
-        type: 'object',
-        required: ['myParameter'],
-        properties: {
-          myParameter: {
-            title: 'An example parameter',
-            description: "This is an example parameter, don't set it to foo",
-            type: 'string',
-          },
-        },
-      },
-    },
-    async handler(ctx) {
-      ctx.logger.info(
-        `Running example template with parameters: ${ctx.input.myParameter}`,
-      );
-
-      if (ctx.input.myParameter === 'foo') {
-        throw new Error(`myParameter cannot be 'foo'`);
-      }
-
-      await new Promise(resolve => setTimeout(resolve, 1000));
-    },
-  });
-}
